softwarearchitektur backend bis auf notifications angepasst
